scripts: management scripts creating and deleting users
These scripts simplify the tasks of creating and deleting users from
keystone/openstack by wrapping the openstack commands required to do so.

The create user shows how to create a user and it's project. It will,
also, create the identity file in ~/openstack-credentials, if it exists.

The delete user script deletes the user and, optinally, the project the
user had assigned in the identity file; at ~/openstack-credentials